Update ApiClient.cs
The reason I asked for the ability to pass in HttpClient is so a single reusable instance with a logging handler could be passed in.

As such, the HttpClient base address should not be set if the client is passed in. This will throw an InvalidOperationException if the client has been used before.

Additionally, the default headers should be set at the request object as they will be different for each instantiation of FingerprintApi -- even though the HttpClient will be the same.

I know you generally don't accept community pull requests but I'd appreciate if you'd consider this.